requirements.txt
flask==3.0.0
flask-cors==4.0.0
networkx==3.2.1
osmnx==2.0.7
geopy==2.4.0
requests==2.31.0
matplotlib>=3.8.0
Pillow>=11.0.0
geopandas>=0.14.0
shapely>=2.0.0
pandas>=2.0.0
pyproj>=3.6.0
rasterio>=1.3.9  # Optional; needed only for test_nlcd_tiff.py
